# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Everise API",
    "version": "1.0.0",
    "description": "The API behind everise.dev. Generated from the routes (src/api)."
  },
  "paths": {
...
    "/api/articles/{id}/comments": {
      "get": {
        "tags": [
          "articles"
        ],
        "summary": "A post's comments, newest first",
        "security": [
          {
            "token": []
          },
          {}
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The comments.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentsResponse"
                }
              }
            }
          },
          "default": {
            "description": "An error; `errors.body` says what went wrong.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "articles"
        ],
        "summary": "Comment on a post",
        "security": [
          {
            "token": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewComment"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The comment.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentResponse"
                }
              }
            }
          },
          "default": {
            "description": "An error; `errors.body` says what went wrong.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/articles/{id}/comments/{commentId}": {
      "delete": {
        "tags": [
          "articles"
        ],
        "summary": "Delete a comment (its author only)",
        "security": [
          {
            "token": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "commentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Done."
          },
          "default": {
            "description": "An error; `errors.body` says what went wrong.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/profiles/{id}": {
      "get": {
        "tags": [
          "profiles"
        ],
        "summary": "A member's profile",
        "security": [
          {
            "token": []
          },
          {}
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The member.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            }
          },
          "default": {
            "description": "An error; `errors.body` says what went wrong.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/profiles/{id}/follow": {
      "post": {
        "tags": [
          "profiles"
        ],
        "summary": "Follow a member",
        "security": [
          {
            "token": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The member.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            }
          },
          "default": {
            "description": "An error; `errors.body` says what went wrong.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "profiles"
        ],
        "summary": "Stop following a member",
        "security": [
          {
            "token": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The member.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            }
          },
          "default": {
            "description": "An error; `errors.body` says what went wrong.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/waking-sands/characters/{id}/picture": {
      "get": {
        "tags": [
          "waking-sands"
        ],
        "summary": "A character's picture",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The picture.",
            "content": {
              "image/png": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "default": {
            "description": "An error; `errors.body` says what went wrong.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/waking-sands/room": {
      "get": {
        "tags": [
          "waking-sands"
        ],
        "summary": "The room: its characters, who's in, and the day's lines",
        "parameters": [],
        "responses": {
          "200": {
            "description": "The room.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SandsRoom"
                }
              }
            }
          },
          "default": {
            "description": "An error; `errors.body` says what went wrong.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/waking-sands/room/characters/{id}": {
      "post": {
        "tags": [
          "waking-sands"
        ],
        "summary": "Bring a character into the room",
        "security": [
          {
            "token": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Who's in the room now.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresentResponse"
                }
              }
            }
          },
          "default": {
            "description": "An error; `errors.body` says what went wrong.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "waking-sands"
        ],
        "summary": "Send a character out of the room",
        "security": [
          {
            "token": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Who's in the room now.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresentResponse"
                }
              }
            }
          },
          "default": {
            "description": "An error; `errors.body` says what went wrong.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/waking-sands/room/lines": {
      "post": {
        "tags": [
          "waking-sands"
        ],
        "summary": "Say something (the characters' answers arrive live)",
        "security": [
          {
            "token": []
          }
        ],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewSandsLine"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The line.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SandsLineResponse"
                }
              }
            }
          },
          "default": {
            "description": "An error; `errors.body` says what went wrong.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "NewSandsLine": {
        "type": "object",
        "properties": {
          "text": {
            "type": "string",
            "minLength": 1,
            "maxLength": 1000
          }
        },
        "required": [
          "text"
        ]
      },
...
      "NewAttachment": {
        "type": "object",
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "image",
              "gif",
              "video"
            ]
          },
          "url": {
            "type": "string",
            "maxLength": 2000,
            "format": "uri"
          },
          "alt": {
            "type": "string",
            "maxLength": 200
          },
          "width": {
            "type": "integer",
            "minimum": 1,
            "maximum": 10000
          },
          "height": {
            "type": "integer",
            "minimum": 1,
            "maximum": 10000
          }
        },
        "required": [
          "kind",
          "url"
        ]
      },
...
      "NewComment": {
        "type": "object",
        "properties": {
          "comment": {
            "type": "object",
            "properties": {
              "body": {
                "type": "string"
              },
              "media": {
                "maxItems": 1,
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/NewAttachment"
                }
              }
            },
            "required": [
              "body"
            ]
          }
        },
        "required": [
          "comment"
        ]
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "errors": {
            "type": "object",
            "properties": {
              "body": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            },
            "required": [
              "body"
            ],
            "additionalProperties": false
          },
          "unconfirmedEmail": {
            "type": "string"
          }
        },
        "required": [
          "errors"
        ],
        "additionalProperties": false
      },
...
      "SandsLine": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "at": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
          },
          "from": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "image": {
            "type": "string"
          },
          "memberId": {
            "type": "string"
          },
          "text": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "at",
          "from",
          "name",
          "text"
        ],
        "additionalProperties": false
      },
      "SandsRoom": {
        "type": "object",
        "properties": {
          "available": {
            "type": "boolean"
          },
          "characters": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "name": {
                  "type": "string"
                },
                "title": {
                  "type": "string"
                },
                "image": {
                  "type": "string"
                }
              },
              "required": [
                "id",
                "name",
                "title"
              ],
              "additionalProperties": false
            }
          },
          "present": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "lines": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SandsLine"
            }
          },
          "limits": {
            "type": "object",
            "properties": {
              "maxPresent": {
                "type": "integer",
                "minimum": -9007199254740991,
                "maximum": 9007199254740991
              },
              "maxLineLength": {
                "type": "integer",
                "minimum": -9007199254740991,
                "maximum": 9007199254740991
              }
            },
            "required": [
              "maxPresent",
              "maxLineLength"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "available",
          "characters",
          "present",
          "lines",
          "limits"
        ],
        "additionalProperties": false
      },
      "PresentResponse": {
        "type": "object",
        "properties": {
          "present": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "present"
        ],
        "additionalProperties": false
      },
      "SandsLineResponse": {
        "type": "object",
        "properties": {
          "line": {
            "$ref": "#/components/schemas/SandsLine"
          }
        },
        "required": [
          "line"
        ],
        "additionalProperties": false
      },
...
      "Profile": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "following": {
            "type": "boolean"
          },
          "bio": {
            "type": [
              "string",
              "null"
            ]
          },
          "image": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "username",
          "following",
          "bio",
          "image"
        ],
        "additionalProperties": false
      },
      "ProfileResponse": {
        "type": "object",
        "properties": {
          "profile": {
            "$ref": "#/components/schemas/Profile"
          }
        },
        "required": [
          "profile"
        ],
        "additionalProperties": false
      },
      "Attachment": {
        "anyOf": [
          {
            "type": "object",
            "properties": {
              "kind": {
                "type": "string",
                "enum": [
                  "image",
                  "gif"
                ]
              },
              "url": {
                "type": "string"
              },
              "alt": {
                "type": "string"
              },
              "width": {
                "type": "integer",
                "minimum": -9007199254740991,
                "maximum": 9007199254740991
              },
              "height": {
                "type": "integer",
                "minimum": -9007199254740991,
                "maximum": 9007199254740991
              }
            },
            "required": [
              "kind",
              "url"
            ],
            "additionalProperties": false
          },
          {
            "type": "object",
            "properties": {
              "kind": {
                "type": "string",
                "const": "video"
              },
              "url": {
                "type": "string"
              },
              "videoId": {
                "type": "string"
              },
              "start": {
                "type": "integer",
                "minimum": -9007199254740991,
                "maximum": 9007199254740991
              },
              "alt": {
                "type": "string"
              }
            },
            "required": [
              "kind",
              "url",
              "videoId"
            ],
            "additionalProperties": false
          }
        ]
      },
...
      "Comment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
          },
          "body": {
            "type": "string"
          },
          "media": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Attachment"
              },
              {
                "type": "null"
              }
            ]
          },
          "author": {
            "$ref": "#/components/schemas/Profile"
          }
        },
        "required": [
          "id",
          "createdAt",
          "updatedAt",
          "body",
          "media",
          "author"
        ],
        "additionalProperties": false
      },
      "CommentResponse": {
        "type": "object",
        "properties": {
          "comment": {
            "$ref": "#/components/schemas/Comment"
          }
        },
        "required": [
          "comment"
        ],
        "additionalProperties": false
      },
      "CommentsResponse": {
        "type": "object",
        "properties": {
          "comments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Comment"
            }
          }
        },
        "required": [
          "comments"
        ],
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "token": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "`Token <jwt>`, or the `token` cookie the site sets."
      }
    }
  }
}